first draft for new splitter
