+++ clinicadl/splitter/__init__.py
@@ -1,0 +1,1 @@
+from .make_splits import make_kfold, make_split
