+++ clinicadl/splitter/make_splits/__init__.py
@@ -1,0 +1,2 @@
+from .kfold import make_kfold
+from .single_split import make_split
