+++ clinicadl/splitter/splitter/__init__.py
@@ -1,0 +1,2 @@
+from .kfold import KFold
+from .single_split import SingleSplit
